backend/requirements.txt
djangorestframework
django-cors-headers
drf-yasg
pytesseract
django-phonenumber-field[phonenumbers]
pillow
